fix: rename tableCellNode to CustomTableCellNode
