+++ packages/lexical-playground/src/App.tsx
@@ -20,8 +20,8 @@ import {SharedAutocompleteContext} from './context/SharedAutocompleteContext';
 import {SharedHistoryContext} from './context/SharedHistoryContext';
 import Editor from './Editor';
 import logo from './images/logo.svg';
+import {CustomTableCellNode} from './nodes/CustomTableCellNode';
 import PlaygroundNodes from './nodes/PlaygroundNodes';
-import {CustomTableCellNode} from './nodes/TableCellNode';
 import DocsPlugin from './plugins/DocsPlugin';
 import PasteLogPlugin from './plugins/PasteLogPlugin';
 import {TableContext} from './plugins/TablePlugin';
