+++ packages/lexical-playground/src/nodes/CustomTableCellNode.tsx
